requirements.txt
diffusers
transformers
torch
pillow
trimesh